// src/app/models/user.ts
export class User {
    s_no?: number;
    first_name?: string;
    last_name?: string;
    address?: string;
    email?: string;
    gender?: string;
    city?: string;
    zip_code?: string;
    contact_number?: string;
    qualification?: string;
    experience?: number;
    hobbies?: string;
    Remarks?: string;
}